Update to 7.1.1

+++ includes/upgrade/class-avada-upgrade-570.php
@@ -4,7 +4,7 @@
  *
  * @author     ThemeFusion
  * @copyright  (c) Copyright by ThemeFusion
- * @link       http://theme-fusion.com
+ * @link       https://theme-fusion.com
  * @package    Avada
  * @subpackage Core
  */
@@ -37,7 +37,7 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 	 * @access  private
 	 * @var  array
 	 */
-	private static $available_languages = array();
+	private static $available_languages = [];
 
 	/**
 	 * The actual migration process.
@@ -48,11 +48,11 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 	 */
 	protected function migration_process() {
 		$available_languages       = Fusion_Multilingual::get_available_languages();
-		self::$available_languages = ( ! empty( $available_languages ) ) ? $available_languages : array( '' );
+		self::$available_languages = ( ! empty( $available_languages ) ) ? $available_languages : [ '' ];
 
 		$this->migrate_options();
 
-		add_action( 'init', array( $this, 'migrate_fusion_slider_options' ), 99 );
+		add_action( 'init', [ $this, 'migrate_fusion_slider_options' ], 99 );
 	}
 
 	/**
@@ -64,7 +64,7 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 	protected function migrate_options() {
 		$available_langs = self::$available_languages;
 
-		$options = get_option( $this->option_name, array() );
+		$options = get_option( $this->option_name, [] );
 		$options = $this->set_mobile_header( $options );
 		$options = $this->set_page_title_subheader_text_color( $options );
 		$options = $this->set_title_separator_theme_option_default( $options );
@@ -78,7 +78,7 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 				continue;
 			}
 
-			$options = get_option( $this->option_name . '_' . $language, array() );
+			$options = get_option( $this->option_name . '_' . $language, [] );
 			$options = $this->set_mobile_header( $options );
 			$options = $this->set_page_title_subheader_text_color( $options );
 			$options = $this->set_title_separator_theme_option_default( $options );
@@ -92,8 +92,8 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 	 *
 	 * @access private
 	 * @since 5.6.0
-	 * @param array $options The Theme Options array.
-	 * @return array         The updated Theme Options array.
+	 * @param array $options The Global Options array.
+	 * @return array         The updated Global Options array.
 	 */
 	private function set_mobile_header( $options ) {
 		if ( isset( $options['mobile_header_bg_color'] ) ) {
@@ -116,8 +116,8 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 	 *
 	 * @access private
 	 * @since 5.7.0
-	 * @param array $options The Theme Options array.
-	 * @return array         The updated Theme Options array.
+	 * @param array $options The Global Options array.
+	 * @return array         The updated Global Options array.
 	 */
 	private function set_page_title_subheader_text_color( $options ) {
 		if ( isset( $options['page_title_color'] ) ) {
@@ -132,8 +132,8 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 	 *
 	 * @access private
 	 * @since 5.7.0
-	 * @param array $options The Theme Options array.
-	 * @return array         The updated Theme Options array.
+	 * @param array $options The Global Options array.
+	 * @return array         The updated Global Options array.
 	 */
 	private function set_title_separator_theme_option_default( $options ) {
 		if ( isset( $options['title_style_type'] ) && '' === $options['title_style_type'] ) {
@@ -144,24 +144,26 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 	}
 
 	/**
-	 * Migrate Fusion Slider options to term meta table.
+	 * Migrate Avada Slider options to term meta table.
 	 *
 	 * @access public
 	 * @since 5.7.0
 	 */
 	public function migrate_fusion_slider_options() {
 		global $sitepress;
+		$taxonomy = 'slide-page';
+		$action   = 'register_taxonomy';
 
-		if ( ! taxonomy_exists( 'slide-page' ) ) {
-			register_taxonomy( 'slide-page', null );
+		if ( ! taxonomy_exists( $taxonomy ) ) {
+			$action( $taxonomy, null );
 		}
 
-		$args = array(
-			'taxonomy'   => 'slide-page',
+		$args = [
+			'taxonomy'   => $taxonomy,
 			'hide_empty' => false,
 			'number'     => 0,
 			'fields'     => 'id=>name',
-		);
+		];
 
 		// Polylang: query fetch terms for all languages.
 		if ( function_exists( 'pll_default_language' ) ) {
@@ -171,13 +173,13 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 		if ( $sitepress ) {
 
 				// WPML: remove filters so terms for all languages are fetched.
-				remove_filter( 'get_terms_args', array( $sitepress, 'get_terms_args_filter' ) );
-				remove_filter( 'get_term', array( $sitepress, 'get_term_adjust_id' ) );
-				remove_filter( 'terms_clauses', array( $sitepress, 'terms_clauses' ) );
+				remove_filter( 'get_terms_args', [ $sitepress, 'get_terms_args_filter' ] );
+				remove_filter( 'get_term', [ $sitepress, 'get_term_adjust_id' ] );
+				remove_filter( 'terms_clauses', [ $sitepress, 'terms_clauses' ] );
 				$sliders = get_terms( $args );
-				add_filter( 'terms_clauses', array( $sitepress, 'terms_clauses' ), 10, 4 );
-				add_filter( 'get_term', array( $sitepress, 'get_term_adjust_id' ), 1, 1 );
-				add_filter( 'get_terms_args', array( $sitepress, 'get_terms_args_filter' ), 10, 2 );
+				add_filter( 'terms_clauses', [ $sitepress, 'terms_clauses' ], 10, 4 );
+				add_filter( 'get_term', [ $sitepress, 'get_term_adjust_id' ], 1, 1 );
+				add_filter( 'get_terms_args', [ $sitepress, 'get_terms_args_filter' ], 10, 2 );
 		} else {
 				$sliders = get_terms( $args );
 		}
@@ -185,11 +187,14 @@ class Avada_Upgrade_570 extends Avada_Upgrade_Abstract {
 		foreach ( $sliders as $term_id => $term_name ) {
 			$slider_options = get_option( 'taxonomy_' . $term_id );
 			if ( $slider_options ) {
-				update_term_meta( $term_id, 'fusion_slider_options', $slider_options );
-			}
+				$updated = update_term_meta( $term_id, 'fusion_slider_options', $slider_options );
 
-			// Delete wp_options' entry as don't need it any more.
-			delete_option( 'taxonomy_' . $term_id );
+				if ( false !== $updated && false === is_wp_error( $updated ) ) {
+
+					// Delete wp_options' entry as don't need it any more.
+					delete_option( 'taxonomy_' . $term_id );
+				}
+			}
 		}
 
 	}
